move commands -> cmd/continuity/commands
