+++ cmd/continuity/commands/verify.go
